demo-app/requirements.txt
flask==2.3.3
Werkzeug==2.3.7
Jinja2==3.1.2
itsdangerous==2.1.2
click==8.1.7
MarkupSafe==2.1.3
requests==2.31.0
urllib3==2.0.5
chardet==5.2.0
idna==3.4
certifi==2023.7.22
SQLAlchemy==2.0.21
python-dotenv==1.0.0 